move api call to one location
