+++ src/layouts/Admin.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Switch, Route, Redirect } from "react-router-dom";
 
 // components
@@ -16,15 +16,34 @@ import Settings from "views/admin/Settings.js";
 import Tables from "views/admin/Tables.js";
 
 export default function Admin() {
+  const [measures, setMeasures] = useState([]);
+
+  useEffect(() => {
+    fetch(`${process.env.REACT_APP_HEDIS_MEASURE_API_URL}measures`)
+      .then(response => response.json())
+      .then(res => {
+        const specialName = "Composite Score";
+        const first = res.find(a => a.name === specialName);
+        const theRestSorted = res.filter(a => a.name !== specialName).sort((a, b) => a.name.localeCompare(b.name));
+        setMeasures([first, ...theRestSorted])
+    })
+  }, []);
+
   return (
     <>
       <Sidebar />
       <div className="relative md:ml-64 bg-gray-200">
         <AdminNavbar />
         {/* Header */}
-        <HeaderStats />
+        <HeaderStats measures={measures} />
         <div className="px-4 md:px-10 mx-auto w-full -m-24">
           <Switch>
+            <Route
+              path='/admin/dashboard'
+              render={(props) => (
+                <Dashboard {...props} measures={measures} />
+              )}
+            />
             <Route path="/admin/dashboard" exact component={Dashboard} />
             <Route path="/admin/maps" exact component={Maps} />
             <Route path="/admin/settings" exact component={Settings} />
